Move query scopes
