+++ src/Query/Scopes/PaymentStatusFilter.php
@@ -1,48 +1,41 @@
 <?php
 
-namespace DoubleThreeDigital\SimpleCommerce\Scopes;
+namespace DoubleThreeDigital\SimpleCommerce\Query\Scopes;
 
 use DoubleThreeDigital\SimpleCommerce\Orders\EntryOrderRepository;
-use DoubleThreeDigital\SimpleCommerce\Orders\OrderStatus;
+use DoubleThreeDigital\SimpleCommerce\Orders\PaymentStatus;
 use DoubleThreeDigital\SimpleCommerce\SimpleCommerce;
 use DoubleThreeDigital\SimpleCommerce\Support\Runway;
 use Statamic\Query\Scopes\Filter;
 
-class OrderStatusFilter extends Filter
+class PaymentStatusFilter extends Filter
 {
     public $pinned = true;
 
-    public static $title = 'Order Status';
+    public static $title = 'Payment Status';
 
     public function fieldItems()
     {
         return [
             'type' => [
                 'type' => 'radio',
-                'options' => collect(OrderStatus::cases())->mapWithKeys(fn ($case) => [
-                    $case->value => __($case->name),
+                'options' => collect(PaymentStatus::cases())->mapWithKeys(fn ($case) => [
+                    $case->value => $case->name,
                 ])->toArray(),
             ],
         ];
     }
 
-    public function autoApply()
-    {
-        return [
-            'type' => 'placed',
-        ];
-    }
-
     public function apply($query, $values)
     {
-        return $query->where('order_status', $values['type']);
+        return $query->where('payment_status', $values['type']);
     }
 
     public function badge($values)
     {
-        $orderStatusLabel = OrderStatus::from($values['type'])->name;
+        $paymentStatusLabel = PaymentStatus::from($values['type'])->name;
 
-        return __('Order Status: :orderStatus', ['orderStatus' => $orderStatusLabel]);
+        return "Payment Status: {$paymentStatusLabel}";
     }
 
     public function visibleTo($key)
